structurizing page-content
