+++ client/src/components/page-content/page-content.css
@@ -1,8 +1,10 @@
+/* UTILITY CLASSES */
 .content {
     display: flex;
-    margin: 0 auto;
+    /* margin: 0 auto; */
 
-    height: 800px;
+    height: 1200px;
+    /* padding: 1em; */
     /* width: 80%; */
 
     flex-direction: column;
@@ -10,56 +12,30 @@
     align-items: center;
 }
 
-/* welcome section of the page (usually what the user sees when landing) */
-.welcome-container {
-    /* position: relative; */
-    height: auto;
-    display: flex;
-    margin: 0 auto;
-    width: 80%;
 
-    padding: 1.7rem 0rem ;
-
-
-    max-width: 100ch;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-
-    font-size: 1.5rem;
-    /* margin: 0; */
-    /* text-align: center; */
-}
-
-/* Description of the company or the services maybe */
-.aboutus-container{
-    display: flex;
-    flex-direction: row;
-    height: 300px;
-    width: 80%;
-    justify-content: space-evenly;
-    align-items: center;
-}
-
-.snowflake-icon_container {
-    display: flex;
-    height: auto;
-    width: auto;
+/* ANIMATION CLASSES*/
+@keyframes snowflakeMovement {
     
-    align-items: center;
-}
-.snowflake-icon img{
-    width: 15rem;
-    height: 15rem;
-}
-
-.description_container {
-    display: flex;
-    height: auto;
-
-    align-items: center;
+    0% {
+        transform: rotate(0.025turn);
+    }
+    
+    50% {
+        transform: rotate(-0.025turn);
+    }
+    100% {
+        transform: rotate(0.025turn);
+    }
 }
 
-.description-abtus {
-    max-width: 60ch;
+.snowflakeAnimation {
+    /* animation: 
+    name
+    duration
+    timingFunction
+    delay
+    iterationCount
+    direction
+    fillMode */
+    animation: snowflakeMovement 10s ease-in-out 0s infinite normal both;
 }
