+++ client/src/components/page-content/smaller-components/Welcome.jsx
@@ -1,0 +1,17 @@
+import './welcome.css'
+
+const Welcome = () => {
+    return(
+    <div className="welcome-container">
+        <div className="welcome-description">
+            asdfasdf, asdfa  asdfa asdfa ,asdfa asdfa ,asdfaasdfaasdfa, asdfa
+            sdafasdfaasdfaasdfa asdfa asdfaasdfa  asdfaasdfaasdfa
+            asdfsadfasdf
+            asdfasdfasdfasdf
+            asdfasdf
+        </div>
+    </div>
+    )
+}
+
+export default Welcome;
