+++ client/src/components/page-content/smaller-components/aboutus.css
@@ -1,0 +1,45 @@
+/* Description of the company or the services maybe */
+.aboutus-container{
+    display: flex;
+    flex-direction: row;
+    height: 300px;
+    width: 80%;
+    justify-content: space-evenly;
+    align-items: center;
+}
+.snowflake-icon_container {
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
+    gap: 10px;
+
+    /* height: auto;
+    width: auto; */
+    /* justify-content: center;
+    align-items: center; */
+}
+.snowflake-icon img{
+    width: 100px;
+    height: 100px;
+
+    /* transition:
+    duration
+    timingFunction
+    delay
+     */
+    transition: transform 0.5s ease 0s;
+}
+.snowflake-icon img:hover {
+    transform: scale(1.1, 1.1);
+}
+
+
+/* DESCRIPTION CONTAINER */
+.description_container {
+    display: flex;
+    height: auto;
+
+    align-items: center;
+}
+.description-abtus {
+    max-width: 60ch;
+}
