+++ client/src/components/page-content/smaller-components/index.js
@@ -1,0 +1,2 @@
+export { default as Aboutus } from "./Aboutus";
+export { default as Welcome } from "./Welcome";
