+++ client/src/components/page-content/smaller-components/welcome.css
@@ -1,0 +1,28 @@
+/* welcome section of the page (usually what the user sees when landing) */
+.welcome-container {
+    position: relative;
+    height: auto;
+    /* width: 100%; */
+    display: flex;
+    /* margin: 0 auto; */
+    width: 80%;
+
+    padding: 1.7rem 0rem ;
+
+
+    /* max-width: 100ch; */
+    /* flex-direction: row; */
+    justify-content: center;
+    align-items: center;
+
+    /* font-size: 1.5rem; */
+    /* margin: 0; */
+    /* text-align: center; */
+}
+.welcome-description {
+    display: flex;
+
+    font-size: 1.5rem;
+    max-width: 100ch;
+    text-align: center;
+}
